+++ src/Sentry/ITransactionContext.cs
@@ -1,9 +1,11 @@
+using Sentry.Protocol;
+
 namespace Sentry;
 
 /// <summary>
 /// Transaction metadata.
 /// </summary>
-public interface ITransactionContext : ISpanContext
+public interface ITransactionContext : ITraceContext
 {
     /// <summary>
     /// Transaction name.
@@ -14,4 +16,9 @@ public interface ITransactionContext : ISpanContext
     /// Whether the parent transaction of this transaction has been sampled.
     /// </summary>
     bool? IsParentSampled { get; }
+
+    /// <summary>
+    /// The source of the transaction name.
+    /// </summary>
+    TransactionNameSource NameSource { get; }
 }